backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
supabase==2.10.0
python-dotenv==1.0.1
pydantic==2.10.3
python-jose[cryptography]==3.3.0

